// src/enums/AuditAction.enum.ts
export enum AuditAction {
  CREATE_ADMIN = 'CREATE_ADMIN',
  EDIT_ADMIN = 'EDIT_ADMIN',
  DELETE_ADMIN = 'DELETE_ADMIN',
  RETRIEVE_ADMINS = 'RETRIEVE_ADMINS',
  RETRIEVE_ADMIN_BY_ID = 'RETRIEVE_ADMIN_BY_ID',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  VERIFY_RESET_OTP = 'VERIFY_RESET_OTP',
  RESET_PASSWORD = 'RESET_PASSWORD',
  USER_LOGIN = 'USER_LOGIN',
  WEBAUTHN_AUTH_OPTIONS = 'WEBAUTHN_AUTH_OPTIONS',
  WEBAUTHN_VERIFY = 'WEBAUTHN_VERIFY',
  CREATE_DEPARTMENT = 'CREATE_DEPARTMENT',
  RETRIEVE_DEPARTMENT = 'RETRIEVE_DEPARTMENT',
  UPDATE_DEPARTMENT = 'UPDATE_DEPARTMENT',
  DELETE_DEPARTMENT = "DELETE_DEPARTMENT",
  RETRIEVE_DEPARTMENT_USERS = 'RETRIEVE_DEPARTMENT_USERS',
  RETRIEVE_DOMAINS = 'RETRIEVE_DOMAINS',
  CREATE_DOMAIN = 'CREATE_DOMAIN',
  RETRIEVE_DOMAIN_BY_ID = 'RETRIEVE_DOMAIN_BY_ID',
  UPDATE_DOMAIN = 'UPDATE_DOMAIN',
  DELETE_DOMAIN = 'DELETE_DOMAIN',
  TOKEN_GENERATION = 'TOKEN_GENERATION',
}